fix: Improve build to include --add-exports when using gradle application bin.
