+++ build.gradle
@@ -26,7 +26,20 @@ dependencies {
 
 application {
     // Define the main class for the application
-    mainClass = 'org.eisopux.diagnostics.json.lsp.Main'
+    mainClass = 'org.eisopux.diagnostics.demo.DemoMain'
+    applicationDefaultJvmArgs = [
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.api=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.util=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.main=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.processing=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.file=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.tree=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.model=ALL-UNNAMED',
+            '--add-exports', 'jdk.compiler/com.sun.tools.javac.comp=ALL-UNNAMED',
+            '--add-opens', 'jdk.compiler/com.sun.tools.javac.api=ALL-UNNAMED',
+            '--add-opens', 'jdk.compiler/com.sun.tools.javac.processing=ALL-UNNAMED'
+    ]
 }
 
 tasks.withType(JavaCompile) { compilationTask ->
